+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K70XQKZYDH25W3KTG334VPA0"
+        "01K71A099F1PQBV64ZFPAV6FVW"
     )
     r"""The id of the resource"""