+++ packages/orq-rc/RELEASES.md
@@ -3199,3 +3199,13 @@ Based on:
 - [python v3.14.0-rc.15] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.15] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.15 - packages/orq-rc
+
+## 2025-10-08 07:03:19
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.636.0 (2.723.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.16] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.16] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.16 - packages/orq-rc